autooooo

+++ sim_p1_v2/parallel/device_utils/device_cavi_update.cu
@@ -25,10 +25,10 @@ __device__ void device_cavi_implementation::device_cavi_point_estimate_update(){
       sum_phi += phi[i][k];
       product_x_phi += x[i] * phi[i][k];
     }
-    s2[k] = 1 / (1/sigma_2 + sum_phi);
+    s2[k] = 1 / (1/double(sigma_2) + sum_phi);
     m[k] = product_x_phi * s2[k];
 
-    elbo += - (product_x_phi * product_x_phi + 1) * s2[k]/(2*sigma_2) + log(s2[k])/2.;
+    elbo += - (product_x_phi * product_x_phi + 1) * s2[k]/(2.*sigma_2) + log(s2[k])/2.;
     for(int i = 0; i < n_samples; i++){
       elbo += phi[i][k] *
         (x[i] * (- x[i]/2. + m[k]) - (s2[k] + m[k]*m[k])/2.);
